Moved RemoteFeedImageMapper to it's own file
